node: cache sexp for `inspect` and co
This is the true fix for a bug we had when running tests in parallel.

Consider the following: You have an array `arr` containing a node, and
you call `arr.non_existing_method`. This should raise an exception and
it does.

The exception message will call the array's `inspect`, and
subsequently all its children's `inspect`.

When the underlying `TSTree*` is still in memory, as in in serial tests,
then calling `ts_node_string`, through `inspect`, is safe because the
message will be shown immediately, therefore `inspect` will be evaluated
before `GC`ing the `Tree` object.

However, when running tests in parallel, using `minitest_parallel_fork`
for example, the exception message is not evaluated directly, and kept
till the end of the test run. When the time comes to show the error
message, the `Tree` object we used to extract the `Node` would have been
`GC`ed already, so calling `ts_node_string` at this point will cause a
`SEGFAULT`.